added conversion from t_INT to INT, now using INT for all HNF/SNF invariants

+++ pari_snf.h
@@ -4,13 +4,14 @@
 #define _PARI_SNF_H      1       //flags that this file has been included
 
 #include <eclib/templates.h>
+#include "int.h"
 
 // Return a list of the Smith Normal Form invariants of a matrix
 // (encoded as vector<vector<int>>)
-vector<long> invariants(const vector<vector<int>>& M);
+vector<INT> invariants(const vector<vector<int>>& M);
 
 // Return list of entries of the HNF of M (exncoded as above),
 // assuming M has size 2 (i.e. represnts a 2xn matrix):
-vector<long> hnf_invariants(const vector<vector<int>>& M);
+vector<INT> hnf_invariants(const vector<vector<int>>& M);
 
 #endif
